diary content structure update

+++ script.js
@@ -2,9 +2,11 @@ const diaryWrapper = document.getElementById("diary-wrapper");
 let deleteBtn = document.createElement("button");
 let keys = Object.keys(localStorage);
 let lsLength = localStorage.length;
+let uniqueFinalDates = [];
 let lsArr = [];
 
 function updateValues() {
+    console.log(keys)
     keys = Object.keys(localStorage);
     getDateObject();
     lsLength = localStorage.length;
@@ -17,7 +19,7 @@ function updateValues() {
 
 function rerenderContent(lsLength) {
     for(let i = 0; i < lsLength; i++) {
-       let element = diaryWrapper.getElementsByClassName("diary-content")[0];
+       let element = diaryWrapper.getElementsByClassName("unique-diary-wrapper")[0];
 
        if (element) {
         element.remove();
@@ -28,86 +30,101 @@ function rerenderContent(lsLength) {
 function showContent(lsLength) {
     updateValues();
     rerenderContent(lsLength);
-    for(let i = 0; i < localStorage.length; i++) {
-        const diary = localStorage.getItem(keys[i]);
-        const parsedDiary = JSON.parse(diary);
-        const a = parsedDiary.untilDate;
+    for(let i = 0; i < uniqueFinalDates.length; i++) {
+        const uniqueDate = uniqueFinalDates[i];
+        console.log(uniqueFinalDates)
+        let diaryEntries = lsArr.filter(obj => obj.untilDate === uniqueDate);
+        console.log(diaryEntries);
+        const uniqueDateWrapper = document.createElement("div");
+        const uniqueDateTxt = uniqueDate;
 
-        const dContent = document.createElement("div");
-        const p = document.createElement("p");
-        const h2 = document.createElement("h2");
-        const time = document.createElement("span");
-        const buttonWrapper = document.createElement("div");
-        const buttons = deleteBtn.cloneNode(true);
-        const diaryText = document.createElement("div");
+        uniqueDateWrapper.setAttribute("id", uniqueDate);
+        uniqueDateWrapper.classList.add("unique-diary-wrapper");
 
-        dContent.classList.add("diary-content")
-        const result = countTime(a);
-        if (result <= 1) {
-            time.style.color = "red";
+        diaryWrapper.append(uniqueDateWrapper); 
+        uniqueDateWrapper.append(uniqueDateTxt);     
+        
+        for(let x = 0; x < diaryEntries.length; x++) {
+            const diaryEntry = diaryEntries[x];
+            console.log(diaryEntry)
+            const dContent = document.createElement("div");
+            const diaryText = document.createElement("div");
+            const p = document.createElement("p");
+            const h2 = document.createElement("h2");
+            const time = document.createElement("span");
+            const buttonWrapper = document.createElement("div");
+            const buttons = deleteBtn.cloneNode(true);
+
+            dContent.classList.add("diary-content")
+            const result = countTime(uniqueDate);
+            console.log(result)
+            if (result <= 1) {
+                time.style.color = "red";
+            } 
+            if (result < 0) {
+                const entriesToDlt = lsArr.filter(obj => obj.untilDate === uniqueDate);
+                for(let y = 0; y < entriesToDlt.length; y++) {
+                    const key = entriesToDlt[y].key;
+                    localStorage.removeItem(key);
+                    showContent();
+                }
+            } 
+
+            buttons.textContent = "Delete";
+            buttons.classList.add("delete-btn");
+            buttons.setAttribute(`id`, diaryEntry.key);
+            buttons.addEventListener("click", function() {
+                const lsLength = localStorage.length;
+                const keyId = buttons.getAttribute("id");
+                localStorage.removeItem(keyId);
+                showContent(lsLength);
+            });
+
+            uniqueDateWrapper.append(dContent)
+            dContent.append(diaryText);
+            dContent.append(buttonWrapper);
+            diaryText.append(h2);
+            diaryText.append(p);
+            diaryText.append(time);
+            h2.innerText = diaryEntry.title;
+            p.innerText = diaryEntry.description;
+            time.innerText = diaryEntry.untilDate;
+            buttonWrapper.append(buttons);
         }
-        console.log(result);
-
-        buttons.textContent = "Delete";
-        buttons.classList.add("delete-btn");
-        buttons.setAttribute(`id`, keys[i]);
-        buttons.addEventListener("click", function() {
-            const lsLength = localStorage.length;
-            const keyId = buttons.getAttribute("id");
-            localStorage.removeItem(keyId);
-            //deleteItems(lsLength);  
-            showContent(lsLength);
-        });
-
-        diaryWrapper.append(dContent)
-        dContent.append(diaryText);
-        dContent.append(buttonWrapper);
-        diaryText.append(h2);
-        diaryText.append(p);
-        diaryText.append(time);
-        h2.innerText = parsedDiary.title;
-        p.innerText = parsedDiary.description;
-        time.innerText = parsedDiary.untilDate;
-        buttonWrapper.append(buttons);
     }
     console.log(localStorage)
 }
 
-function countTime(a) {
+function countTime(uniqueDate) {
     const date = new Date;
     const year = date.getFullYear();
     const month = (date.getMonth() + 1).toString().padStart(2, '0');
     const day = (date.getDate()).toString().padStart(2, '0');
     const currentDate = `${year}${month}${day}`;
-    const finalDate = a.replace(/-/g, "")
+    const finalDate = uniqueDate.replace(/-/g, "")
     const result = parseInt(finalDate) - parseInt(currentDate);
-
     return result;
 }
 
 function getDateObject() {
     lsArr.length = 0;
+    uniqueFinalDates.length = 0;
     for (let i = 0; i < localStorage.length; i++) {
         const ls = localStorage.getItem(keys[i]);
         const parsedLs = JSON.parse(ls);
         lsArr.push(parsedLs)
     }
     const finalDates = lsArr.filter(obj => parseInt(obj.untilDate) > 1  ).map(obj => obj.untilDate);
-    const uniqueFinalDates = [...new Set(finalDates)]
+    uniqueFinalDates = [...new Set(finalDates)]
     console.log(uniqueFinalDates)
 }
 
 document.getElementById("form").addEventListener("submit", function(event) {
     event.preventDefault();
+    keys.length = 0;
     const title = document.getElementById("titleInput");
     const description = document.getElementById("descInput");
     const inputDate = document.getElementById("dateInput");
-    const diary = {
-        title: title.value,
-        description: description.value,
-        untilDate: inputDate.value
-    }
-    //const lsLength = localStorage.length;
     const time = new Date;
     const date = {
         year: time.getFullYear(),
@@ -120,63 +137,28 @@ document.getElementById("form").addEventListener("submit", function(event) {
     };
     const dateKey = `${date.year}${date.month}${date.day}${date.hours}${date.minutes}${date.seconds}${date.milSeconds}`;
 
-    localStorage.setItem(dateKey, JSON.stringify(diary));
+    const diary = {
+        title: title.value,
+        description: description.value,
+        untilDate: inputDate.value,
+        key: dateKey
+    }
+
+    if(diary.untilDate.replace(/-/g, "") - parseInt(`${date.year}${date.month}${date.day}`) < 0) {
+        title.value = "";
+        description.value = "";
+        inputDate.value = "";
+        
+    } else {
+        localStorage.setItem(dateKey, JSON.stringify(diary));
+    } 
+
     title.value = "";
     description.value = "";
     inputDate.value = "";
-    //deleteItems(lsLength);
     showContent(lsLength);
 });
 
-
-
 document.addEventListener('DOMContentLoaded', function() {
-    //updateValues();
     showContent();
 });
-
-/**
- for(let i = 0; i < localStorage.length; i++) {
-        const diary = localStorage.getItem(keys[i]);
-        const parsedDiary = JSON.parse(diary);
-        const a = parsedDiary.untilDate;
-
-        const dContent = document.createElement("div");
-        const p = document.createElement("p");
-        const h2 = document.createElement("h2");
-        const time = document.createElement("span");
-        const buttonWrapper = document.createElement("div");
-        const buttons = deleteBtn.cloneNode(true);
-        const diaryText = document.createElement("div");
-
-        dContent.classList.add("diary-content")
-        const result = countTime(a);
-        if (result <= 1) {
-            time.style.color = "red";
-        }
-        console.log(result);
-
-        buttons.textContent = "Delete";
-        buttons.classList.add("delete-btn");
-        buttons.setAttribute(`id`, keys[i]);
-        buttons.addEventListener("click", function() {
-            const lsLength = localStorage.length;
-            const keyId = buttons.getAttribute("id");
-            localStorage.removeItem(keyId);
-            //deleteItems(lsLength);  
-            showItems(lsLength);
-        });
-
-        diaryWrapper.append(dContent)
-        dContent.append(diaryText);
-        dContent.append(buttonWrapper);
-        diaryText.append(h2);
-        diaryText.append(p);
-        diaryText.append(time);
-        h2.innerText = parsedDiary.title;
-        p.innerText = parsedDiary.description;
-        time.innerText = parsedDiary.untilDate;
-        buttonWrapper.append(buttons);
-    }
-    console.log(localStorage)
- */
